took the conversion to the provider layer, and updated tests
